api 提交

+++ dataanalysis/src/main/java/com/aizhixin/cloud/dataanalysis/analysis/dto/TeachingAchievementDTO.java
@@ -15,11 +15,11 @@ public class TeachingAchievementDTO {
     @ApiModelProperty(value = "学生人数", required = false)
     private int studentsNum;
     @ApiModelProperty(value = "平均GPA", required = false)
-    private float averageGPA;
+    private double averageGPA;
     @ApiModelProperty(value = "不及格人数", required = false)
     private int FailNum;
     @ApiModelProperty(value = "开设课程数量", required = false)
     private int coursesNum;
     @ApiModelProperty(value = "课程平均得分", required = false)
-    private float coursesAVGScore;
+    private double coursesAVGScore;
 }
